% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimateData.R
\name{getClimateData}
\alias{getClimateData}
\title{getClimateData: Download daily Daymet gridded climate data}
\usage{
getClimateData(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = c(2006:2023),
  filepath = NA,
  export = FALSE,
  silent = TRUE
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Vector of years to download Daymet data for. Earliest available year is 1980. Latest is currently 2023.}

\item{filepath}{Quoted path to save files to. If not specified, will save to working directory.}

\item{export}{Logical. If TRUE, will export a CSV of the compiled Daymet data with a date stamp. Must supply
a filepath to write output to. If FALSE (Default), will only return results to R environment.}

\item{silent}{Logical. If TRUE (Default), won't show Daymet comments in console. If FALSE, will print Daymet comments
as sign of progress (and working) in console.}
}
\value{
Data frame of Daymet daily climate data for each specified site.
}
\description{
This function downloads daily climate data for each selected NETN water monitoring site
based on its lat/long coordinates, and binds each site's data into a single dataframe. Final dataframe
can also be written to disk (export = T). If downloading for all sites and multiple years, function
may be slow.
}
\examples{
\dontrun{
# RUN FIRST
library(waterNETN)
importData()

#++++++ ADD EXAMPLES +++++++

}

}
