% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimDaymet.R
\name{getClimDaymet}
\alias{getClimDaymet}
\title{getClimDaymet: Download daily Daymet gridded climate data directly from REST}
\usage{
getClimDaymet(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = c(2006:2023),
  active = TRUE,
  weather_station = FALSE
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Vector of years to download Daymet data. Earliest available year is 1980. Latest is
currently 12/31/2023.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites that have been monitored.}

\item{weather_station}{Logical. If TRUE, will return Daymet data for coordinates of nearest weather station to a park.
If FALSE (default), returns Daymet data for water monitoring sites only. This argument facilitates calculating decadal
averages to compare with current-year weather station data.}
}
\value{
Data frame of Daymet daily climate data for each specified site.
}
\description{
This function downloads daily gridded climate data directly from Daymet REST for
each selected NETN water monitoring site based on its lat/long coordinates, and binds each
site's data into a single dataframe. If downloading for all sites and multiple years, function
may be slow. The returned data frame includes Day length (dayl in s/day), Precipitation (prcp
in mm/day), Shortwave radiation (srad in W/m2), Snow water equivalent (swe in kg/m2),
Maximum air temperature (tmax in C), Minimum air temperature (tmin in C), and Water vapor
pressure (vp in Pascals). More details on metrics can be found online:
https://daymet.ornl.gov/overview.html > Parameters, Parameter abbreviations, Units and Descriptions.
Note that occasionally you're unable to connect to the server, and will receive an error message
when that happens.
}
\examples{
\dontrun{
# RUN FIRST
library(waterNETN)
importData()

# download for MORR 2023 only
morr <- getClimDaymet(park = "MORR", years = 2023)

# download for ACAD lakes from 1980:2023
acad_lakes <- getClimDaymet(park = "ACAD", site_type = "lake", years = 1980:2023)

}

}
