% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimWStat.R
\name{getClimWStat}
\alias{getClimWStat}
\title{getClimWStat: Download daily data from nearest weather station}
\usage{
getClimWStat(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = c(2006:2023),
  filepath = NA,
  export = FALSE
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Vector of years to download weather station data for, will start with 01/01/year and end with 12/31/year.
Note that not all weather stations have complete a complete period of record from 2006 to current.}

\item{filepath}{Quoted path to save files to. If not specified, will save to working directory.}

\item{export}{Logical. If TRUE, will export a CSV of the compiled weather data with a date stamp. Must supply
a filepath to write output to. If FALSE (Default), will only return results to R environment.}
}
\value{
Data frame of weather station daily climate data for each specified site.
}
\description{
This function downloads daily weather data from data.rcc-acis.org from the nearest
weather station with the longest period of record dating through 2024 for each selected NETN water
monitoring site based on its lat/long coordinates. Data include daily total precipitation (mm),
max. temperature (C), min temperature (C) and binds each site's data into a single dataframe. Final dataframe
can also be written to disk (export = T). If downloading for multiple sites and multiple years, function
may be slow, particularly ACAD. Note that the McFarland Hill weather station precip. data is not available
through typical weather station sources. Instead, hourly precip data from NADP are downloaded and summarized
to daily value. ACAD NADP Precip data only go back to 2008.
}
\examples{
\dontrun{
# RUN FIRST
library(waterNETN)
importData()

# get weather data for ROVA sites from 2020-2023
rova <- getClimWStat(park = "ROVA", years = 2020:2023)

# get weather data for the Pogue (if can't remember the sitecode)
mabi <- getClimWStat(park = "MABI", site_type = "lake")

# get weather data for 2023 for all lakes in ACAD (this will be slow)
acad_lakes <- getClimWStat(park = "ACAD", years = 2023, site_type = 'lake')

# get weather data for Aunt Betty Inlet for 2022 and 2023
abin <- getClimWStat(site = "ACABIN", years = 2022:2023)

}

}
