% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getSites}
\alias{getSites}
\title{getSites: query NETN water data by site}
\usage{
getSites(park = "all", site = "all", site_type = c("all", "lake", "stream"))
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}
}
\value{
Data frame of site info
}
\description{
Queries NETN water data by site. Returned data frame can include lakes or streams, but only includes
columns that the Site_Lake and Site_Stream views have in common. For full site data, use the site-specific getSites()
functions (i.e., getSitesLakes()). This function is a useful building block within other package functions. Must
have water quality views imported in R session.
}
\examples{
\dontrun{
importData()

# get site info for all sites in MABI
mabi <- getSites(park = "MABI")

# get site info for all streams in ACAD
ACAD_streams <- getSites(park = 'ACAD', site_type = 'stream')

# get site info East Primrose Brook in MORR
east_prim <- getSites(site = "MORRSA")
}
}
