% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSitesLake.R
\name{getSitesLake}
\alias{getSitesLake}
\title{getSitesLake: query NETN lakes by site}
\usage{
getSitesLake(park = "all", site = "all", output = c("short", "verbose"))
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{output}{Specify if you want all fields returned (output = "verbose") or just the most important fields (output = "short"; default.)}
}
\value{
Data frame of site info for lakes
}
\description{
Queries NETN lakes by site.
}
\examples{
\dontrun{
importData()

# get site info for all lakes in lowerNETN
lnetn <- c("MABI", "MIMA", "MORR", "ROVA", "SAGA", "SAIR", "SARA", "WEFA")
lnetn_lks <- getSitesLake(park = lnetn)

# get site info for all lakes in ACAD
ACAD_lakes <- getSitesLake(park = 'ACAD')

}
}
