% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSondeInSitu.R
\name{getSondeInSitu}
\alias{getSondeInSitu}
\title{getSondeInSitu: query NETN water chemistry Sonde data}
\usage{
getSondeInSitu(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  parameter = "all",
  QC_type = "0",
  sample_depth = "surface",
  output = c("short", "verbose")
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Numeric. Years to query. Accepted values start at 2006.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{parameter}{Specify the chemical parameter(s) to return. Note if additional parameters are added to the Chemistry view, there will be additional
to the views, they will be added as accepted values in this function. Current accepted values are:.
c("Temp_C", "SpCond_uScm", "DOsat_pct", "DOsatLoc_pct", "DO_mgL", "pH", "pHmV",
"Turbidity_FNU", "ChlA_RFU", "ChlA_ugL", "BP_mmHg")}

\item{QC_type}{Specify QC type to return. Must be quoted.
\describe{
\item{"all"}{Include all QC types.}
\item{"0"}{Environmental. Default. Indicates a real non-QC sample.}
\item{"900"}{QC Replicate}
\item{"899"}{Pre-deployment 100 percent sat check}
\item{"999"}{Post-deployment 100 percent sat check}
}}

\item{sample_depth}{Filter on sample depth. If "all", returns all sample depths. If "surface" (Default),
only returns the median value of samples collected <= 2m from the surface. SampleDepth_m is also the median
sample depth of samples collected within 2m of the surface.}

\item{output}{Specify if you want all fields returned (output = "verbose") or just the most important fields (output = "short"; default.)}
}
\value{
Data frame of Sonde chemistry data in long form.
}
\description{
Queries NETN water chemistry data with the Sonde in the field by site, event, and parameter.
If pulling all sites, parameters, years, etc., function may take a second or two to run.
}
\examples{
\dontrun{
importData()

# get Sonde data for all sites and parameters in MABI from 2021-2023 for non-QAQC samples
mabi <- getSondeInSitu(park = "MABI", years = 2021:2023)

# get data for all DO parameters in MIMA from 2006-2023 for non-QAQC samples
params <- c("DOsat_pct", "DOsatLoc_pct", "DO_mgL")
period <- 2006:2023
mima_n <- getSondeInSitu(park = "MIMA", years = period, parameter = params)

# get Temp data for surface measurements only in Jordan Pond in ACAD for non-QAQC samples
ACAD_lake <- getSondeInSitu(site = 'ACJORD', parameter = "Temp_C", sample_depth = "surface")

# get pH for lower NETN parks from May to Oct for QAQC and non-QAQC samples
lnetn <- c("MABI", "MIMA", "MORR", "ROVA", "SAGA", "SAIR", "SARA", "WEFA")
lnetn_ph <- getSondeInSitu(park = lnetn, param = "pH", months = 5:10, QC_type = 'all')
}
}
