% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStreamObs.R
\name{getStreamObs}
\alias{getStreamObs}
\title{getStreamObs: query NETN stream observations}
\usage{
getStreamObs(
  park = "all",
  site = "all",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  output = c("short", "verbose")
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{years}{Numeric. Years to query. Accepted values start at 2006.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{output}{Specify if you want all fields returned (output = "verbose") or just the most important fields (output = "short"; default.)}
}
\value{
Data frame of stream observations
}
\description{
Queries NETN stream observations by park, site, year and month.
}
\examples{
\dontrun{
importData()

# get stream observations for Pogue Stream all years
mabi <- getStreamObs(park = "MABI")

# get observations for all streams in ACAD May 2023
ACAD_streams <- getStreamObs(park = 'ACAD', years = 2023, months = 5)

}
}
