% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimComps.R
\name{plotClimComps}
\alias{plotClimComps}
\title{plotClimComps: Plot climate comparisons}
\usage{
plotClimComps(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 1:12,
  active = TRUE,
  layers = "lines",
  data_type = "daymet",
  averages = "norm80",
  parameter = "tmean",
  plot_title = TRUE,
  palette = "viridis",
  color_rev = FALSE,
  legend_position = "right",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site.
Defaults to "all". If multiple sites sites selected, will facet on site.}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Numeric. Years to plot separately. Accepted values start at 2006.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites.
If FALSE, returns all sites that have been monitored.}

\item{layers}{Options are "points", "lines", or both for annual averages. By default, only lines will plot.}

\item{data_type}{Specify Daymet ("daymet"; default), or weather station ("wstn"). Note that Daymet
data aren't available immediately, whereas weather station data may be available within days of current day.
When 'wstn' is selected, the averages are based on Daymet data at the coordinates of the weather station,
because the closest weather stations to most parks do not have a period or record prior to 2000.}

\item{averages}{Specify averages to plot. By default, the 30-year normal from 1980-2009 plots. Options include
one of the following:
\describe{
\item{"norm80"}{Plots the 30-year norm from 1980 - 2009}
\item{"norm90"}{Plots the 30-year norm from 1990 - 2019}
\item{"decades"}{Plots 1980 - 1989, 1990 - 1999, 2000 - 2009, and 2010 - 2019 averages}
}}

\item{parameter}{Specify the monthly averaged parameter to plot. Acceptable values are
\describe{
\item{"tmean"}{Plot mean temperature comparisons (in C).}
\item{"tmax"}{Plot max temperature comparisons (in C).}
\item{"tmin"}{Plot min temperature comparisons (in C).}
\item{"ppt"}{Plot precipitation comparisons (in mm).}
}}

\item{plot_title}{Logical. If TRUE (default) prints site name at top of figure. If FALSE,
does not print site name. Only enabled when one site is selected.}

\item{palette}{Color palette for plots. Options currently are 'viridis' (yellow - green - blue),
or create your own by specifying 2 or more colors in quotes, like palette = c("red", "yellow", "blue"),
or palette = c("yellow", "blue"). Hexcodes work too. If only 1 year is specified, only need to
specify 1 color}

\item{color_rev}{Reverse the order of the color pallet. Only enabled for palette = 'viridis'.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position =
"none" (Default). Other options are "top", "bottom", "left", "right".}

\item{...}{Additional arguments relevant to \code{sumClimAvgs()} or \code{sumClimMonthly()}}
}
\value{
Returns a ggplot object of specified climate trends
}
\description{
This function compares decadal (1980 - 1989, ..., 2010 - 2019) average
monthly climate variables with user-specified years to provide an idea of how extreme
or normal a given month in a year is. This function only works with Daymet data,
so is not immediately available for the current date.
}
\examples{
\dontrun{

# Plot mean monthly temp for the Pogue in MABI for 2019:2023 and all months with red-blue color palette
plotClimComps(site = "MABIPA", years = 2019:2023, parameter = "temp_mean", palette = c('red', 'blue'))

# Same as above, but with points too
plotClimComps(site = "MABIPA", years = 2019:2023, parameter = "temp_mean", palette = c('red', 'blue'),
layers = c('points', 'lines'))

# Plot max monthly temp for Eagle Lake and Jordan Pond for 2019:2023 and all months
plotClimComps(site = c("ACEAGL", "ACJORD"), years = 2019:2023, parameter = "temp_max",
palette = c('red', 'blue'), layers = 'lines')

# Plot total monthly precip for Jordan Pond for past 10 years using a blue color scheme
plotClimComps(site = "ACJORD", years = 2013:2023, parameter = 'precip',
palette = c("#75C5FF", "#3563DD", "#323969"))

# Plot total monthly precip for latest year using weather station data
plotClimComps(site = "ACJORD", years = 2024, parameter = "precip", data_type = 'wstn')

}

}
