% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimRel.R
\name{plotClimRel}
\alias{plotClimRel}
\title{plotClimRel: Plot climate data relative to average value}
\usage{
plotClimRel(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 1:12,
  active = TRUE,
  averages = "norm80",
  parameter = "tmean",
  plot_title = TRUE,
  palette = "viridis",
  color_rev = FALSE,
  legend_position = "right",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site.
Defaults to "all". If multiple sites sites selected, will facet on site.}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Numeric. Years to plot separately. Accepted values start at 2006.If multiple years
specified, will facet results on year. Note that selecting many years and/or many sites
may be slow to return results.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites.
If FALSE, returns all sites that have been monitored.}

\item{averages}{Specify averages to plot. By default, the 30-year normal from 1980-2009 plots. Options include
one of the following:
\describe{
\item{"norm80"}{Plots the 30-year norm from 1980 - 2009}
\item{"norm90"}{Plots the 30-year norm from 1990 - 2019}
}}

\item{parameter}{Specify the monthly averaged parameter to plot. Acceptable values are
\describe{
\item{"temp"}{Plot all temperature comparisons (in C).}
\item{"tmean"}{Plot mean temperature comparisons (in C).}
\item{"tmax"}{Plot max temperature comparisons (in C).}
\item{"tmin"}{Plot min temperature comparisons (in C).}
\item{"ppt"}{Plot precipitation comparisons (in mm).}
}}

\item{plot_title}{Logical. If TRUE (default) prints site name at top of figure. If FALSE,
does not print site name. Only enabled when one site is selected.}

\item{palette}{Color palette for plots. Options currently are 'viridis' (yellow - green - blue),
or create your own by specifying 2 or more colors in quotes, like palette = c("red", "yellow", "blue"),
or palette = c("yellow", "blue"). Hexcodes work too. If only 1 year is specified, only need to
specify 1 color}

\item{color_rev}{Reverse the order of the color pallet. Only enabled for palette = 'viridis'.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position =
"none" (Default). Other options are "top", "bottom", "left", "right".}

\item{...}{Additional arguments relevant to \code{sumClimAvgs()} or \code{sumClimMonthly()}}
}
\value{
Returns a ggplot object of specified climate trends
}
\description{
This function plots a given year compared against either the 1980 - 2009
normal or the 1990 - 2019 normal. The 30-year normals are derived from Daymet data. The
annual comparisons are derived from the nearest weather station to each park.
}
\examples{
\dontrun{

#++++ ADD EXAMPLES ++++

}

}
