% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLakeProfile.R
\name{plotLakeProfile}
\alias{plotLakeProfile}
\title{plotLakeProfile: Plots smoothed trend}
\usage{
plotLakeProfile(
  park = "all",
  site = "all",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  active = TRUE,
  parameter = NA,
  depth_type = "raw",
  color_theme = "spectral",
  color_rev = FALSE,
  plot_title = TRUE,
  plot_thermocline = TRUE,
  legend_position = "right",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{years}{Numeric. Years to query. Accepted values start at 2006 for depth_type = "surface". For depth_type = "elev", accepted
values start at 2013, as water level data prior to 2013 are not available in the data package.}

\item{months}{Numeric. Months to query by number. This function is restricted to months 5:10. X-axis may plot weird if specifying
anything but 5:10 for months.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites that have been monitored.}

\item{parameter}{Specify the parameter to return (can only choose 1 per function call). Current accepted values are:.
c("Temp_C", "SpCond_uScm", "DOsat_pct", "DOsatLoc_pct", "DO_mgL", "pH", "pHmV",
"Turbidity_FNU", "ChlA_RFU", "ChlA_ugL", "BP_mmHg").}

\item{depth_type}{Specify whether to plot sample depth relative to elevation of surface water (depth_type = "elev") or
depth, with each sample starting at 0 regardless of level of the lake surface (depth_type = "raw"; default).}

\item{color_theme}{Divering color palette for plots. Options currently are 'spectral' (default), 'ryb' (red - yellow - blue),
'rb' (red - blue), and 'viridis' (yellow - green - blue) (see https://ggplot2-book.org/scales-colour for more info).}

\item{color_rev}{Reverse the order of the color pallete. For example change ryb from red - yellow - blue to blue - yellow -red.}

\item{plot_title}{Logical. If TRUE (default) prints site name at top of figure. If FALSE, does not print site name. Only enabled when
one site is selected.}

\item{plot_thermocline}{Logical. If TRUE (default) plots the depth of the thermocline, calculated by rLakeAnalyzer as the depth/elevation
within the water column where the temperature gradient is the steepest and indicates where the upper waters are typically not mixing with
deeper waters. Only plots where at least 5 depth measurements for temperature have been collected (may want to increase this threshold).
Note that in the rare cases that multiple sampling events occur within a month, only the first is plotted. If no thermocline is detected,
as defined by \code{rLakeAnalyzer::thermo.depth()}, no points are plotted.}

\item{legend_position}{Specify location of legend (default is 'right'). To turn legend off, use legend_position = "none". Other
options are "top", "bottom", "left", "right".}

\item{...}{Additional arguments relevant to \code{getSondeInSitu()} or \code{getWaterLevel()}}
}
\value{
Returns a panel of hydrographs during the growing season for each year
in the data frame.
}
\description{
This function produces a heatmap filtered on park (mostly for ACAD), site, year, month, Sonde in situ
parameter and either sample relative to the surface or relative to surface elevation. The y-axis is 1m bins. If multiple
samples are taken within the same 1-m bin, the median value is used. The width of the profile columns reflects the
number of days between sample events, centered on the day sampled. Shorter intervals between events result in a narrower bar.
You can only specify one parameter at a time. If multiple sites or years are selected, plots will be faceted on those factors.
Keep options limited for best plotting. The option to plot relative to surface elevation uses the water level data and
datum elevation to convert sample depth to 1-m binned elevations. This allows you to see how the water column is shifting over time.
If you specify a lake x year x parameter combination that doesn't exist (e.g., a year a lake isn't sampled), the
function will return an error message instead of an empty plot.
}
\examples{
\dontrun{



}

}
