% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterPlot.R
\name{plotScatterPlot}
\alias{plotScatterPlot}
\title{plotScatterPlot: Plot scatterplot of 2 variables}
\usage{
plotScatterPlot(
  park = "all",
  site = "all",
  site_type = "all",
  event_type = "VS",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  active = TRUE,
  parameters = NA,
  sample_depth = "surface",
  layers = c("points", "smooth"),
  palette = "viridis",
  facet_site = TRUE,
  facet_scales = "free_y",
  numcol = 2,
  span = 0.3,
  legend_position = "none",
  gridlines = "none",
  ...
)
}
\arguments{
\item{park}{Character or character vector. Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Character or character vector. Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Character. Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{event_type}{Character. Select the event type,can only choose one option. Valid inputs:
\describe{
\item{"all"}{All possible sampling events.}
\item{"VS"}{Default. NETN Vital Signs monitoring events, which includes Projects named 'NETN_LS' and 'NETN+ACID'.}
\item{"acid"}{Acidification monitoring events in Acadia.}
\item{"misc"}{Miscellaneous sampling events.}
}}

\item{years}{Numeric. Years to query. Accepted values start at 2006.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns
all sites.}

\item{parameters}{Specify the two parameters to plot. The first parameter will be treated as the response (y).
The second parameter will be treated as the explanatory (x) variable. Note that censored values can not be included
in the scatterplot. Current accepted values are:\cr

Chemistry
\describe{
\item{"ANC_ueqL}{Acid neutralizing capacity, in micrograms per liter}
\item{"AppColor_PCU"}{Apparent  color, in platinum cobalt units}
\item{"ChlA_ugL"}{Chlorophyll a, in micrograms per liter}
\item{"Cl_ueqL"}{Chloride, in microequivalents per liter}
\item{"DOC_mgL"}{Dissolved organic carbon, in miligrams per liter}
\item{"NH3_mgL"}{Ammonia, in miligrams per liter}
\item{"NO2_mgL"}{Nitrite, in miligrams per liter}
\item{"NO2+NO3_mgL"}{Nitrite and nitrate, in miligrams per liter}
\item{"NO3_ueqL"}{Nitrate, in micrograms per liter}
\item{"pH_Lab"}{pH, as determined by lab}
\item{"PO4_ugL"}{Phosphate, in micrograms per liter}
\item{"SO4_ueqL"}{Sulfate, in micrograms per liter}
\item{"TN_mgL"}{Total nitrogen, in micrograms per liter}
\item{"TP_ugL"}{Total phosphorus, in micrograms per liter}
\item{"TotDissN_mgL"}{Total dissolved nitrogen, in micrograms per liter}
\item{"TotDissP_ugL"}{Total dissolved phosphorus, in micrograms per liter}}

Sonde
\describe{
\item{"Temp_C"}{Temperature of the water sample in degrees celsius.}
\item{"Temp_F"}{Temperature of the water sample in degrees fahrenheit.}
\item{"SpCond_uScm"}{Specific conductivity of the water sample measured in microsiemens per liter.}
\item{"DOsat_pct"}{Dissolved oxygen of the water sample measured in percent saturation.}
\item{"DOsatLoc_pct"}{Post-deployment calibration checks.}
\item{"DO_mgL"}{Dissolved oxygen of the water sample measured in milligrams per liter.}
\item{"pH"}{Quantitative measure of the acidity or basicity of the water sample in pH standard units.}
\item{"pHmV"}{Quantitative measure of the acidity or basicity of the water sample in millvolts.}
\item{"Turbidity_FNU"}{Turbidity of water sample measured in formazin nephelometric units.}
\item{"ChlA_EXO_RFU"}{Chlorophyll a (measure of algae and cyanobacteria) of the water sample in milligrams per liter.}
\item{"ChlA_EXO_ugL"}{Chlorophyll a (measure of algae and cyanobacteria) of the water sample in micrograms per liter.}
\item{"BP_mmHg"}{Barometric pressure, measured in millimeters of mercury.}
}

Other
\describe{
\item{"SDepth_m"}{Secchi disk depth, measured in meters.}
\item{"Discharge_cfs"}{Total discharge; units are cubic feet per second (ft^3/s).}
\item{"PenetrationRatio"}{Specific conductivity of the water sample measured in microsiemens per liter.}
\item{"WaterLevel_Feet"}{Value calculated by adding gage reading to datum elevation (decimal feet).}
\item{"WaterLevel_m"}{Value calculated by adding gage reading to datum elevation (meters).}
}}

\item{sample_depth}{Character. Filter on sample depth. If "all" (Default), returns all sample depths. If "surface",
only returns the median value of samples collected <= 2m from the surface. SampleDepth_m is also the median sample depth of samples collected within 2m of the surface. Note that for the Penetration Ratio parameter,
all sample depths are plotted. Plotting all depths may return a funky plot for other parameters.}

\item{layers}{Character. Options are "points" and "smooth". By default, both will plot. If "smooth" specified, will plot a loess
smoothed line. See "span" for more details. If only points specified, will return a scatterplot.}

\item{palette}{Character. Theme to plot points and lines. Options include 'viridis' (Default- ranges of blue,
green and yellow), magma (yellow, red, purple), plasma (brighter version of magma), turbo (rainbow), RColorBrewer palettes,
or specify a vector of colors manually. If fewer colors than parameters are specified, they will be
ramped to generate enough colors.}

\item{facet_site}{Logical. If TRUE (default), will facet on site if multiple sites specified. If FALSE, will plot all sites
on the same figure. Only enabled when multiple sites specified.}

\item{facet_scales}{Character. Specify whether facet axes should be fixed (all the same; default) or "free_y", "free_x" or "free" (both).}

\item{numcol}{Numeric. Specify number of columns in the facet wrap, which is only enabled when facet_site = T. Default is 2.}

\item{span}{Numeric. Determines how smoothed the line will be for smooth = TRUE. Default is 0.3. Higher spans (up to 1)
cause more smoothing. Span can range from 0 to 1.}

\item{legend_position}{Character. Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{gridlines}{Character. Specify whether to add gridlines or not. Options are: "none" (Default), "grid_y", "grid_x", "both".}

\item{...}{Additional arguments relevant to \code{getChemistry()} or \code{getSondeInSitu()}}
}
\value{
Returns a ggplot scatterplot object.
}
\description{
This function produces points or loess smoothed lines of 2 variables, filtered on park, site, year,
month, and 2 parameters. Works with lab chemistry, Sonde in situ, discharge, secchi depth, water level, and
light penetration ratio. If multiple sites are specified, they will be plotted on the same figure, unless facet_site = T.
Note that if you specify a site and parameter combination that doesn't exist (e.g., a stream site and a parameter
only collected in lakes), the function will return an error message instead of an empty plot. Censored values are
not permitted in this function.
}
\examples{

# Plot Temp vs DO for ROVA all years on same figure
plotScatterPlot(park = "ROVA", parameters = c("DO_mgL", "Temp_C"),
  palette = 'viridis', facet_site = F, legend_position = "bottom")

# Plot Secchi depth vs. surface DOC in Eagle Lake, Jordon Pond, Echo Lake, and Witch Hole Pond
plotScatterPlot(site = c("ACEAGL", "ACJORD", "ACWHOL", "ACECHO"), parameters = c("SDepth_m", "DOC_mgL"),
  span = 0.9, facet_site = F, legend_position = 'bottom')

# Plot Secchi depth vs. surface TN in Eagle Lake, Jordon Pond, and Witch Hole Pond
plotScatterPlot(site = c("ACEAGL", "ACJORD", "ACWHOL"), parameters = c("SDepth_m", "TN_mgL"),
  span = 0.9, facet_site = F, legend_position = 'bottom')

# Plot smoothed discharge vs. specific conductance for the Pogue Brook using span of 0.9.
plotScatterPlot(site = "MABISA", parameters = c("SpCond_uScm", "Discharge_cfs"), span = 0.9, palette = c("forestgreen"))

# Plot smoothed discharge vs. specific conductance for SARA streams using span of 0.9.
plotScatterPlot(park = "SARA", parameters = c("SpCond_uScm", "Discharge_cfs"), span = 0.9, facet_site = F,
legend_position = 'bottom')

# Same as above, but faceted by site.
plotScatterPlot(park = "SARA", parameters = c("SpCond_uScm", "Discharge_cfs"), span = 0.9, facet_site = T)

# Plot TN vs discharge in SARA streams
plotScatterPlot(park = "SARA", parameters = c("TN_mgL", "Discharge_cfs"), span = 0.9, facet_site = F)


}
