% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrendClimate.R
\name{plotTrendClimate}
\alias{plotTrendClimate}
\title{plotTrendClimate: Plots smoothed trend}
\usage{
plotTrendClimate(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 1:12,
  active = TRUE,
  parameter = NA,
  legend_position = "none",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Numeric. Years to query. Accepted values start at 2006.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites that have been monitored.}

\item{parameter}{Specify the parameter(s) to plot. Acceptable values are
\describe{
\item{"all"}{Plot all climate variables}
\item{"dm_ppt_mm"}{Daymet monthly total precipitation in mm.}
\item{"dm_tmax_C"}{Daymet monthly average maximum temperature in C.}
\item{"dm_tmin_C"}{Daymet monthly average minimum temperature in C.}
\item{"dm_tmean_C"}{Daymet monthly average temperature in C.}
\item{"dm_PEThar"}{Daymet potential evapotranspiration using Hargreaves equation}
\item{"dm_BAL"}{Daymet water balance, calculated as precip - PET}
\item{"dm_SPEI01"}{Drought index: Standarized Precipitation Evapotranspiration Index at the 1-month level. }
\item{"dm_SPEI03"}{Drought index: Standarized Precipitation Evapotranspiration Index at the 3-month level. }
\item{"ws_ppt_mm"}{Weather station monthly total precipitation in mm.}
\item{"ws_tmax_C"}{Weather station monthly average maximum temperature in C.}
\item{"ws_tmin_C"}{Weather station monthly average minimum temperature in C.}
\item{"ws_tmean_C"}{Weather station monthly average temperature in C.}
}}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{...}{Additional arguments relevant to \code{getChemistry()} or \code{getSondeInSitu()}}

\item{layers}{Options are "points" and "lines". By default, both will plot.}

\item{color_theme}{Theme to plot points and lines. Options currently are 'viridis' (Default- ranges of blue, green and yellow),
or from RColorBrewer: 'set1', 'dark2', or 'accent' (see https://ggplot2-book.org/scales-colour).}

\item{threshold}{Logical. If TRUE (Default), will plot a dashed (upper) or dotted (lower) line if a water quality threshold exists for that
parameter and site. If FALSE, no threshold line will be plotted.}

\item{smooth}{Logical. If TRUE (Default), will plot a loess smoothed line. If FALSE, will plot actual line. Only
plots if layers argument includes 'lines'.}

\item{span}{Numeric. Determines how smoothed the line will be for smooth = TRUE. Default is 0.3. Higher spans (up to 1)
cause more smoothing. Span can range from 0 to 1.}
}
\value{
Returns a panel of hydrographs during the growing season for each year
in the data frame.
}
\description{
This function produces a line or smoothed trend plot filtered on park, site, year, month, and
climate parameter. If multiple sites are specified, they will be plotted on the same figure. If multiple
parameters are specified, they will be plotted on separate figures.
}
\examples{
\dontrun{



}

}
