% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterBands.R
\name{plotWaterBands}
\alias{plotWaterBands}
\title{plotWaterBands: compare current vs. historic values}
\usage{
plotWaterBands(
  park = "all",
  site = "all",
  site_type = "all",
  event_type = "VS",
  year_current = format(Sys.Date(), "\%Y"),
  years_historic = seq(2006, as.numeric(format(Sys.Date(), "\%Y")) - 1, 1),
  months = 5:10,
  active = TRUE,
  parameter = NA,
  include_censored = FALSE,
  sample_depth = c("surface", "all"),
  threshold = TRUE,
  legend_position = "none",
  gridlines = "none",
  facet_scales = "fixed",
  plotly = FALSE,
  ...
)
}
\arguments{
\item{park}{Character or character vector. Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Character or character vector. Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Character. Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{event_type}{Select the event type, can only choose one option. Valid inputs:
\describe{
\item{"all"}{All possible sampling events.}
\item{"VS"}{Default. NETN Vital Signs monitoring events, which includes projects named 'NETN_LS' and 'NETN+ACID'.}
\item{"acid"}{Acidification monitoring events in Acadia.}
\item{"misc"}{Miscellaneous sampling events.}
}}

\item{year_current}{Numeric. Year that will be plotted separately. Must be 4 digits.}

\item{years_historic}{Number range. Years to include in historic range calculations.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites.}

\item{parameter}{Character. Specify the parameter to return. Can only work with 1 parameter at a time. Current accepted values are:\cr
Chemistry
\describe{
\item{"ANC_ueqL}{Acid neutralizing capacity, in micrograms per liter}
\item{"AppColor_PCU"}{Apparent  color, in platinum cobalt units}
\item{"ChlA_ugL"}{Chlorophyll a, in micrograms per liter}
\item{"Cl_ueqL"}{Chloride, in microequivalents per liter}
\item{"DOC_mgL"}{Dissolved organic carbon, in miligrams per liter}
\item{"NH3_mgL"}{Ammonia, in miligrams per liter}
\item{"NO2_mgL"}{Nitrite, in miligrams per liter}
\item{"NO2+NO3_mgL"}{Nitrite and nitrate, in miligrams per liter}
\item{"NO3_ueqL"}{Nitrate, in micrograms per liter}
\item{"pH_Lab"}{pH, as determined by lab}
\item{"PO4_ugL"}{Phosphate, in micrograms per liter}
\item{"SO4_ueqL"}{Sulfate, in micrograms per liter}
\item{"TN_mgL"}{Total nitrogen, in micrograms per liter}
\item{"TP_ugL"}{Total phosphorus, in micrograms per liter}
\item{"TotDissN_mgL"}{Total dissolved nitrogen, in micrograms per liter}
\item{"TotDissP_ugL"}{Total dissolved phosphorus, in micrograms per liter}}

Sonde
\describe{
\item{"Temp_C"}{Temperature of the water sample in degrees celsius.}
\item{"Temp_F"}{Temperature of the water sample in degrees fahrenheit.}
\item{"SpCond_uScm"}{Specific conductivity of the water sample measured in microsiemens per liter.}
\item{"DOsat_pct"}{Dissolved oxygen of the water sample measured in percent saturation.}
\item{"DOsatLoc_pct"}{Post-deployment calibration checks.}
\item{"DO_mgL"}{Dissolved oxygen of the water sample measured in milligrams per liter.}
\item{"pH"}{Quantitative measure of the acidity or basicity of the water sample in pH standard units.}
\item{"pHmV"}{Quantitative measure of the acidity or basicity of the water sample in millvolts.}
\item{"Turbidity_FNU"}{Turbidity of water sample measured in formazin nephelometric units.}
\item{"ChlA_EXO_RFU"}{Chlorophyll a (measure of algae and cyanobacteria) of the water sample in milligrams per liter.}
\item{"ChlA_EXO_ugL"}{Chlorophyll a (measure of algae and cyanobacteria) of the water sample in micrograms per liter.}
\item{"BP_mmHg"}{Barometric pressure, measured in millimeters of mercury.}
}

Other
\describe{
\item{"SDepth_m"}{Secchi disk depth, measured in meters.}
\item{"Discharge_cfs"}{Total discharge; units are cubic feet per second (ft^3/s).}
\item{"PenetrationRatio"}{Specific conductivity of the water sample measured in microsiemens per liter.}
\item{"WaterLevel_Feet"}{Value calculated by adding gage reading to datum elevation (decimal feet).}
\item{"WaterLevel_m"}{Value calculated by adding gage reading to datum elevation (meters).}
}}

\item{include_censored}{Logical. If TRUE, the value column includes non-censored and censored values
using the MDL/MRL/UQL values in the parameter flags. If the Flag column is not NA, that indicates
the value is a censored value. If FALSE (Default), only non-censored values are returned in the value column.}

\item{sample_depth}{Character. Filter on sample depth. If "all" (Default), returns all sample depths. If "surface",
only returns the median value of samples collected <= 2m from the surface. SampleDepth_m is also the median
sample depth of samples collected within 2m of the surface. Note that for the Penetration Ratio parameter,
all sample depths are plotted. Plotting all depths may return a funky plot for other parameters.}

\item{threshold}{Logical. If TRUE (Default), will plot a dashed (upper) or dotted (lower) line if a water quality
threshold exists for that parameter and site. If FALSE, no threshold line will be plotted.}

\item{legend_position}{Character. Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{gridlines}{Character. Specify whether to add gridlines or not. Options are: "none" (Default), "grid_y", "grid_x", "both"}

\item{facet_scales}{Character. Specify whether facet axes should be "fixed" (all the same) or "free_y", "free_x" or "free" (both).}

\item{plotly}{Logical. If TRUE, returns a plotly object. If FALSE (default), returns a ggplot2 object.}

\item{...}{Additional arguments relevant to \code{getChemistry()} or \code{getSondeInSitu()}}
}
\value{
Returns a ggplot object of specified current vs historic values.
}
\description{
This function produces a plot that summarizes the range of historic data compared with current measurements.
Historic measurements are displayed as min-max values ever previously recorded (outermost band), upper and lower 95\%
distribution and middle 50\% distribution (inner quartiles) of values previously recorded (inner bands). The line
represents the median value.
}
\examples{
\dontrun{
# Plot pH in Jordan Pond for 2023, with gridlines on the y-axis
plotWaterBands(site = "ACJORD", year_curr = 2023, years_historic = 2006:2022,
parameter = "pH", legend_position = 'right', gridlines = 'grid_y')

# Plot TN in Jordan Pond for 2023, including censored and gridlines on both x and y axes
plotWaterBands(site = "ACJORD", year_curr = 2023, years_historic = 2006:2022,
parameter = "TN_mgL", legend_position = 'right', include_censored = T, gridlines = "both")

#' #Same as above, but drop threshold lines and add gridlines on x-axis
plotWaterBands(site = "ACJORD", year_curr = 2023, years_historic = 2006:2022,
parameter = "TN_mgL", legend_position = 'right', include_censored = T, threshold = F, gridlines = "grid_x")

# Plot TN in Kroma Kill in SARA in 2023 with no gridlines (Default)
plotWaterBands(site = "SARASA", year_curr = 2023, years_historic = 2006:2022, parameter = "TN",
legend_position = 'right')
}

}
