% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumClimAvgs.R
\name{sumClimAvgs}
\alias{sumClimAvgs}
\title{sumClimAvgs: Summarize monthly climate norms}
\usage{
sumClimAvgs(
  park = "all",
  site = "all",
  site_type = "all",
  months = 1:12,
  data_type = "daymet",
  active = TRUE,
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest
way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying
particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{data_type}{Specify Daymet ("daymet"), weather station ("wstn"), or both ('all'; default). Note that Daymet
data aren't available immediately, whereas weather station data may be available within days of current day.
When 'wstn' is selected, the averages are based on Daymet data at the coordinates of the weather station,
because the closest weather stations to most parks do not have a period or record prior to 2000.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites.
If FALSE, returns all sites that have been monitored.}

\item{...}{Additional arguments relevant to \code{getSites()} or \code{getClimDaymet()}}
}
\value{
Data frame of averaged monthly climate variables.
}
\description{
This function calculates monthly decadal and 30-year norms from
daily Daymet data by site. For each decade from 1980 to 2020 and for 30-year
spans through 2020, this function returns monthly averages for total precipitation
in mm (ppt_mm), average maximum temperature (tmax_C),average minimum temperature
(tmin_C), and average temperature (tmean_C) based on Daymet gridded climate data.
The 30-year norms from 1980 - 2009 and 1990 - 2019 are designated by _30yr_1980
for 1980-2009, and _30yr_1990 for 1990 - 2020. Averages by decade are designated
by _d1980 for 1980-1989, _d1990 for 1990 - 1999, etc. Note that Daymet data are
downloaded from an online web service and can be slow if many sites selected, and
occasionally the server is down that will result in an error message.
}
\examples{
\dontrun{
# RUN FIRST
library(waterNETN)
importData()

# calculate averages for the Pogue
pogue <- sumClimAvgs(site = "MABIPA")

# Calculate for all active sites in MIMA
mima <- sumClimAvgs(park = "MIMA", start_year = 1990)

}

}
