% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumClimMonthly.R
\name{sumClimMonthly}
\alias{sumClimMonthly}
\title{sumClimMonthly: Summarize monthly climate statistics}
\usage{
sumClimMonthly(
  park = "all",
  site = "all",
  site_type = "all",
  active = TRUE,
  years = c(2006:2023),
  data_type = "all",
  months = 1:12,
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest
way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying
particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites.
If FALSE, returns all sites that have been monitored.}

\item{years}{Vector of years to download weather station data for, will start with 01/01/year and end with 12/31/year.
Note that not all weather stations have complete a complete period of record from 2006 to current.}

\item{data_type}{Specify Daymet ("daymet"), weather station ("wstn"), or both ('all'; default). Note that Daymet
data aren't available immediately, whereas weather station data may be available within days of current day.
When Daymet data are requested, results also return 1-month and 3-month SPEI, a drought index.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}
}
\value{
Data frame of monthly statistics for selected parameters.
}
\description{
This function calculates monthly statistics from daily Daymet and/or
weather station data by site. Statistics returned are monthly total precipitation
in mm (ppt_mm), average maximum temperature (tmax_C), average minimum temperature (tmin_C),
and average temperature (tmean_C). When Daymet data are requested, results also return a
solar radiation (srad in mjm2), potential evapotranspiration using Hargreaves (PEThar),
water balance (BAL), 1- month and 3-month Standardized Precipitation Evapotranspiration Index (SPEI),
a drought index that takes into account precipitation, temperature, evapotranspiration,
and solar radiation. The 3-month SPEI will be NA for the first 2 months in the year span,
because it's a 3-month average. Daymet vs. weather station metrics are indicated by whether
they start with dm_ or ws_. Note that occasionally you're unable to connect to the Daymet
or rcc-acis server and will receive and error message when that happens.
}
\examples{
\dontrun{
# RUN FIRST
library(waterNETN)
importData()

# sum 2006 - 2023 stats for Pogue including daymet and weather station
pog_stats <- sumClimMonthly(site = "MABIPA", years = 2006:2023)

# sum 2023 weather station stats for all active sites in MIMA
mima_stats <- sumClimMonthly(park = "MIMA", years = 2023, data_type = 'wstn')

}

}
