% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_WQ.R
\name{theme_WQ}
\alias{theme_WQ}
\title{theme_WQ: custom ggplot2 theme for waterNETN}
\usage{
theme_WQ()
}
\value{
This function must be used in conjunction with a ggplot object
}
\description{
This function customizes a theme for plotting NETN water data, including removing the
default panel grids from ggplot2 figures.
}
\examples{
{

 # See plotTrend() function example

}
}
