% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDischarge.R
\name{getDischarge}
\alias{getDischarge}
\title{getDischarge: query NETN water discharge data}
\usage{
getDischarge(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  method = "all",
  rating = c("all", "E", "G", "F", "P"),
  output = c("short", "verbose")
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Numeric. Years to query. Accepted values start at 2006.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{method}{Query data by discharge method. Accepted values are c("ACAD Flowtracker", "ACAD Pygmy",
"Flume", "LNETN Flowtracker", "LNETN Pygmy", "No Measurement", "Rating curve estimate", "Timed float",
"USGS Gage", "Visual estimate", "Volumetric")}

\item{rating}{Filter on measurement rating.
\describe{
\item{"E"}{Excellent}
\item{"G"}{Good}
\item{"F"}{Fair}
\item{"P"}{Poor}
}}

\item{output}{Specify if you want all fields returned (output = "verbose") or just the most important fields (output = "short"; default.)}
}
\value{
Data frame of Discharge data.
}
\description{
Queries NETN water discharge data by site, year, month
}
