% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSecchi.R
\name{getSecchi}
\alias{getSecchi}
\title{getSecchi: query NETN Secchi data}
\usage{
getSecchi(
  park = "all",
  site = "all",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  observer_type = c("all", "first", "second"),
  output = c("short", "verbose")
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{years}{Numeric. Years to query. Accepted values start at 2006.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{observer_type}{Return data from all observers, or only return first observer.
Accepted values are c("all", "first", "second"), with "all" being the default.}

\item{output}{Specify if you want all fields returned (output = "verbose") or just the most important fields (output = "short"; default.)}
}
\value{
Data frame of Secchi data in long form (ie observers stacked).
}
\description{
Queries NETN Secchi data by site, year, month, and observer type. Data are stacked by
observer. Only records with observations are returned. Only runs for site_type = 'lake'.
}
\examples{
\dontrun{
importData()

# get Secchi depth the Pogue from 2021-2023, first observer only
mabi <- getSecchi(sitecode = "MABIPA", years = 2021:2023, observer_type = "first")

# get secchi for all ACAD lakes sampled in July for all observers
ACAD_lake <- getSecchi(park = 'ACAD', months = 7)

}
}
