% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData: Imports NETN water data package}
\usage{
importData(path = NA, new_env = TRUE)
}
\arguments{
\item{path}{Quoted path where data package csvs live. Note that currently any csv in
the path provided will import. We will tighten this up once we make the data package into a zip file.}

\item{new_env}{Logical. Specifies which environment to store views in. If \code{TRUE}(Default), stores
views in VIEWS_WQ environment. If \code{FALSE}, stores views in global environment}
}
\value{
Assigns water csvs to specified environment
}
\description{
This function imports flat files as csvs in the NETN water data package.
Each view is added to a VIEWS_WQ environment in your workspace, or to your global
environment based on whether new_env = TRUE or FALSE.
}
\examples{
\dontrun{
# Import tables using default settings
importData(path = "C:/NETN/Water/data")

# Import tables into global environment
importData(path = "C:/NETN/Water/data", new_env = FALSE)
}

}
