% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimTrend.R
\name{plotClimTrend}
\alias{plotClimTrend}
\title{plotClimTrend: Plot climate trends}
\usage{
plotClimTrend(
  park = "all",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 1:12,
  layers = c("points", "lines"),
  parameter = NA,
  facet_park = FALSE,
  facet_param = TRUE,
  palette = "viridis",
  span = 0.3,
  plot_se = FALSE,
  legend_position = "none",
  gridlines = "none"
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Numeric. Years to query. Accepted values start at 1895.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12.
If specifying new months not yet included in NETN_clim_annual dataset, will
download months that are available from NOAA.}

\item{layers}{Options are "points", "lines", and "smooth". By default, both points and lines will plot. The lines option connects
each monthly value with a linear line. If you include lines and smooth, the loess-smoothed line will also plot.}

\item{parameter}{Specify the parameter(s) to plot. Acceptable values are
\describe{
\item{"all"}{Plot all climate variables}
\item{"ppt"}{Monthly total precipitation in mm.}
\item{"tmax"}{Monthly average maximum temperature in C.}
\item{"tmin}{Monthly average minimum temperature in C.}
\item{"tmean}{Monthly average temperature in C.}
}}

\item{facet_park}{Logical. If TRUE, plots sites on separate facets (ie figures). If FALSE (Default),
plots all sites on the same figure. This is only enabled if multiple sites are chosen.}

\item{facet_param}{Logical. If TRUE (Default), plots parameters on separate facets. If FALSE, plots
all parameters on the same figure. Note that results will be funky if selected parameters have different units (e.g., temp and precip).}

\item{palette}{Theme to plot points and lines. Options currently are 'viridis' (Default- ranges of blue, green and yellow), or discrete palettes from RColorBrewer. Common options are "Set1", "Set2", "Dark2", "Accent".
Run RColorBrewer::display.brewer.all(type = 'qual') to see full set of options.}

\item{span}{Numeric. Determines how smoothed the line will be for layers = 'smooth'. Default is 0.3. Higher spans (up to 1)
cause more smoothing (straighter lines). Smaller spans are wavier. Span can range from 0 to 1. Span of 1 is linear.}

\item{plot_se}{Logical. If TRUE, will plot a standard error ribbon. If FALSE (Default), will not plot a ribbon.
Only enabled if layers = "smooth" is specified.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}
}
\value{
Returns a ggplot object of specified climate trends
}
\description{
This function produces a line or smoothed trend plot filtered on park, year, month, and
climate parameter. If multiple parks are specified, they can either be plotted on the same figure or separate
figures. If multiple parameters are specified, they will be plotted on separate figures.
}
\examples{
\dontrun{

# Plot total monthly precip for 2006:2023 and all months, without smoothing
plotClimTrend(park = "MABI", years = 2006:2023, parameter = "ppt", smooth = F)

# Plot avg temp in MABI from 1895:2024, without points and with smoothed line and span 0.1.
plotClimTrend(park = "MABI", years = 1895:2024, layers = 'lines', parameter = "tmean", span = 0.1)

# Plot monthly mean temperature for MABI and SARA from 2006:2024, with smoothed line, span 0.7,
and only sample months using the Dark2 color palette.
plotClimTrend(park = c("MABI", "SARA"), years = 2006:2024,
              parameter = "tmean", span = 0.7, months = 5:10, palette = "Dark2")

}

}
