% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLakeProfile.R
\name{plotLakeProfile}
\alias{plotLakeProfile}
\title{plotLakeProfile: Plots smoothed trend}
\usage{
plotLakeProfile(
  park = "ACAD",
  site = "all",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  active = TRUE,
  parameter = NA,
  depth_type = "raw",
  palette = "Spectral",
  color_rev = FALSE,
  plot_title = TRUE,
  plot_thermocline = TRUE,
  legend_position = "right",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Because ACAD lakes are binned by 1m intervals,
and LNETN are binned by 0.25m intervals, 'all' is not an option. You either must specify "ACAD" (default), "LNETN" or a LNETN park
with profile data. Valid inputs:
\describe{
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{years}{Numeric. Years to query. Accepted values start at 2006 for depth_type = "surface". For depth_type = "elev", accepted
values start at 2013, as water level data prior to 2013 are not available in the data package.}

\item{months}{Numeric. Months to query by number. This function is restricted to months 5:10. X-axis may plot weird if specifying
anything but 5:10 for months.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites that have been monitored.}

\item{parameter}{Specify the parameter to return (can only choose 1 per function call). Current
accepted values are:
c("Temp_C", "SpCond_uScm", "DOsat_pct", "DOsatLoc_pct", "DO_mgL", "pH", "pHmV",
"Turbidity_FNU", "ChlA_RFU", "ChlA_ugL", "BP_mmHg").}

\item{depth_type}{Specify whether to plot sample depth relative to elevation of surface water
(depth_type = "elev") or depth, with each sample starting at 0 regardless of level of the lake
surface (depth_type = "raw"; default).}

\item{palette}{Diverging color palette for plots. Options currently are 'viridis'
(yellow - green - blue), 'mako' (light blue grading to black), or any built-in continuous color palette
available in RColorBrewer. Run RColorBrewer::display.brewer.all() to see the diverging color
palettes. Common palettes include "Blues", "BuGn", "RdPu", "Spectral", "RdYlBu", "RdBu", "PiYg".
See https://ggplot2-book.org/scales-colour for more info.}

\item{color_rev}{Reverse the order of the color pallete. For example change RdYlBu from red - yellow - blue
to blue - yellow -red.}

\item{plot_title}{Logical. If TRUE (default) prints site name at top of figure. If FALSE,
does not print site name. Only enabled when one site is selected.}

\item{plot_thermocline}{Logical. If TRUE (default) plots the depth of the thermocline, calculated by
rLakeAnalyzer as the depth/elevation within the water column where the temperature gradient is the steepest
and indicates where the upper waters are typically not mixing with deeper waters. Only plots where at least
5 depth measurements for temperature have been collected (may want to increase this threshold). Note that in
the rare cases that multiple sampling events occur within a month, both are plotted, but will have a thinner
column than other samples that are more spread out. If no thermocline is detected, as defined by
\code{rLakeAnalyzer::thermo.depth()}, no points are plotted.}

\item{legend_position}{Specify location of legend (default is 'right'). To turn legend off, use legend_position = "none". Other
options are "top", "bottom", "left", "right".}

\item{...}{Additional arguments relevant to \code{getSondeInSitu()} or \code{getWaterLevel()}}
}
\value{
Returns a panel of hydrographs during the growing season for each year
in the data frame.
}
\description{
This function produces a heatmap filtered on park (mostly for ACAD), site, year, month, Sonde in situ
parameter and either sample relative to the surface or relative to surface elevation. The y-axis is 1m bins for ACAD
and 0.25m bins for LNETN. If multiple samples are taken within the same bin, the median value is used. The width of
the profile columns reflects the number of days between sample events, centered on the day sampled. Shorter intervals
between events result in a narrower bar. You can only specify one parameter at a time. If multiple sites or years are
selected, plots will be faceted on those factors. Keep options limited for best plotting. The option to plot relative
to surface elevation uses the water level data and datum elevation to convert sample depth to 1-m or 0.25m binned elevations.
This allows you to see how the water column is shifting over time. If you specify a lake x year x parameter combination
that doesn't exist (e.g., a year a lake isn't sampled), the function will return an error message instead of an empty
plot.
}
\examples{
\dontrun{

# Plot temperature for Upper Hadlock for years 2013 - 2023 corrected by elevation with thermocline
# plotted as black lines.
plotLakeProfile(site = "ACUHAD", parameter = "Temp_C", depth_type = 'elev',
                years = 2013:2023)

# Plot temperature for Eagle Lake for years 2006 - 2023 with raw sample depth.
# Note that we can go back to 2006 because we're using raw sample depth instead of elevation.
plotLakeProfile(site = "ACEAGL", parameter = "Temp_C", depth_type = 'raw',
               years = 2006:2023)

# Same plot as above, but with no plot title or thermocline.
plotLakeProfile(site = "ACEAGL", parameter = "Temp_C", depth_type = 'raw',
                years = 2006:2023, plot_title = FALSE, plot_thermocline = F)
# Plot temperature for all ACAD lakes sampled in 2023 and raw sample depth.
lakes23 <- c("ACBUBL", "ACEAGL", "ACECHO", "ACJORD", "ACLONG", "ACROUN",
             "ACSEAL", "ACUBRK", "ACUHAD", "ACWHOL")
plotLakeProfile(park = "ACAD", site = lakes23, parameter = "Temp_C",
               depth_type = 'raw', years = 2023)

# Plot DO all ACAD lakes sampled in 2023 and raw sample depth, using reversed ryb palette.
Temperature thermocline is also included as black lines.
plotLakeProfile(park = "ACAD", site = lakes23,
               parameter = "DOsat_pct", depth_type = 'raw', years = 2023,
               palette = 'RdYlBu', color_rev = TRUE)
}

}
