% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrend.R
\name{plotTrend}
\alias{plotTrend}
\title{plotTrend: Plots smoothed trend}
\usage{
plotTrend(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  active = TRUE,
  parameter = NA,
  include_censored = FALSE,
  sample_depth = c("surface", "all"),
  layers = c("points", "lines"),
  palette = "viridis",
  threshold = TRUE,
  smooth = TRUE,
  numcol = 2,
  span = 0.3,
  legend_position = "none",
  gridlines = "none",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{years}{Numeric. Years to query. Accepted values start at 2006.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites that have been monitored.}

\item{parameter}{Specify the parameter(s) to return. Current accepted values are:.
chemistry: c("ANC_ueqL", "AppColor", "AppColor_PCU", "ChlA_ugL", "Cl_ueqL",
"DOC_mgL", "NH3_mgL", "NO2_mgL", "NO2+NO3_mgL", "NO3_ueqL", "pH_Lab", "PO4_ugL", "SO4_ueqL",
"TN_mgL", "TotDissN_mgL", "TotDissP_ugL", "TP_ugL")
sonde: c("Temp_C", "Temp_F", "SpCond_uScm", "DOsat_pct", "DOsatLoc_pct", "DO_mgL", "pH", "pHmV",
"Turbidity_FNU", "ChlA_RFU", "ChlA_ugL", "BP_mmHg").
other: c("SDepth_m", "Discharge_cfs", "PenetrationRatio", "WaterLevel_Feet", "WaterLevel_m").
Note that "all" is not an accepted value, because there are too many to plot.}

\item{include_censored}{Logical. If TRUE, the value column includes non-censored and censored values
using the MDL/MRL/UQL values in the parameter flags. If the Flag column is not NA, that indicates
the value is a censored value. If FALSE (Default), only non-censored values are returned in the value column.}

\item{sample_depth}{Filter on sample depth. If "all" (Default), returns all sample depths. If "surface",
only returns the median value of samples collected <= 2m from the surface. SampleDepth_m is also the median
sample depth of samples collected within 2m of the surface. Note that for the Penetration Ratio parameter,
all sample depths are plotted. Plotting all depths may return a funky plot for other parameters.}

\item{layers}{Options are "points" and "lines". By default, both will plot.}

\item{palette}{Theme to plot points and lines. Options currently are 'viridis' (Default- ranges of blue, green and yellow), or discrete palettes from RColorBrewer. Common options are "Set1", "Set2", "Dark2", "Accent".
Run RColorBrewer::display.brewer.all(type = 'qual') to see full set of options.}

\item{threshold}{Logical. If TRUE (Default), will plot a dashed (upper) or dotted (lower) line if a water quality threshold exists for that
parameter and site. If FALSE, no threshold line will be plotted.}

\item{smooth}{Logical. If TRUE (Default), will plot a loess smoothed line. If FALSE, will plot actual line. Only
plots if layers argument includes 'lines'.}

\item{numcol}{Specify number of columns in the facet wrap, which is only enabled when either multiple years
are specified or multiple parks. Default is 2.}

\item{span}{Numeric. Determines how smoothed the line will be for smooth = TRUE. Default is 0.3. Higher spans (up to 1)
cause more smoothing. Span can range from 0 to 1.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}

\item{...}{Additional arguments relevant to \code{getChemistry()} or \code{getSondeInSitu()}}
}
\value{
Returns a ggplot object of specified parameter trends.
}
\description{
This function produces a line or smoothed trend plot filtered on park, site, year, month, and parameter.
Works with both lab chemistry data and Sonde in situ data. If multiple sites are specified, they will be plotted
on the same figure. If multiple parameters are specified, they will be plotted on separate figures. Note that
if you specify a stream and parameter combination that doesn't exist (e.g., a stream site and a parameter only
collected in lakes), the function will return an error message instead of an empty plot.
}
\examples{
\dontrun{

# Plot smoothed surface pH for Eagle Lake for past 3 years using default span of 0.3 and by default not including the legend.
plotTrend(site = "ACEAGL", parameter = "pH", palette = 'Dark2', years = 2021:2023) + theme_WQ()

# Plot smoothed surface pH for Eagle Lake for all years, removing the legend and using span of 0.75.
plotTrend(site = "ACEAGL", parameter = "pH", span = 0.75)

# Plot smoothed Secchi Depth in Jordan Pond for all years, including the legend, different color palette, and using span of 0.75.
plotTrend(site = "ACJORD", parameter = "SDepth_m", span = 0.75, palette = 'Set1')

# Plot smoothed surface pH for active SARA streams over all years with 0.6 span.
plotTrend(park = "SARA", site = c("SARASA", "SARASC", "SARASD"), site_type = "stream", parameter = "pH",
          legend_position = "right", span = 0.6)

# Plot smoothed surface SO4 for all MIMA streams over all years with 0.6 span
plotTrend(park = "MIMA", site_type = "stream",
          parameter = "SO4_ueqL", legend_position = "right", span = 0.6)

# Plot non-smoothed surface of multiple Sonde parameters for all MIMA streams over all years with 0.6 span.
params <- c("Temp_F", "SpCond_uScm", "DOsat_pct", "pH")
plotTrend(park = "MIMA", site_type = "stream",
          parameter = params, legend_position = "right", span = 0.6)

}

}
