% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterBands.R
\name{plotWaterBands}
\alias{plotWaterBands}
\title{plotWaterBands: compare current vs. historic values}
\usage{
plotWaterBands(
  park = "all",
  site = "all",
  site_type = "all",
  year_current = format(Sys.Date(), "\%Y"),
  years_historic = seq(2006, as.numeric(format(Sys.Date(), "\%Y")) - 1, 1),
  months = 5:10,
  active = TRUE,
  parameter = NA,
  include_censored = FALSE,
  sample_depth = c("surface", "all"),
  threshold = TRUE,
  legend_position = "none",
  gridlines = "none",
  facet_scales = "fixed",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{year_current}{Year that will be plotted separately. Must be numeric and 4 digits.}

\item{years_historic}{Years to include in historic range calculations.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites that have been monitored.}

\item{parameter}{Specify the parameter to return. Can only work with 1 parameter at a time. Current accepted values are:.
chemistry: c("ANC_ueqL", "AppColor", "AppColor_PCU", "ChlA_ugL", "Cl_ueqL",
"DOC_mgL", "NH3_mgL", "NO2_mgL", "NO2+NO3_mgL", "NO3_ueqL", "pH_Lab", "PO4_ugL", "SO4_ueqL",
"TN_mgL", "TotDissN_mgL", "TotDissP_ugL", "TP_ugL")
sonde: c("Temp_C", "Temp_F", "SpCond_uScm", "DOsat_pct", "DOsatLoc_pct", "DO_mgL", "pH", "pHmV",
"Turbidity_FNU", "ChlA_RFU", "ChlA_ugL", "BP_mmHg").
other: c("SDepth_m", "Discharge_cfs", "PenetrationRatio", "WaterLevel_Feet", "WaterLevel_m").
Note that "all" is not an accepted value, because there are too many to plot.}

\item{include_censored}{Logical. If TRUE, the value column includes non-censored and censored values
using the MDL/MRL/UQL values in the parameter flags. If the Flag column is not NA, that indicates
the value is a censored value. If FALSE (Default), only non-censored values are returned in the value column.}

\item{sample_depth}{Filter on sample depth. If "all" (Default), returns all sample depths. If "surface",
only returns the median value of samples collected <= 2m from the surface. SampleDepth_m is also the median
sample depth of samples collected within 2m of the surface. Note that for the Penetration Ratio parameter,
all sample depths are plotted. Plotting all depths may return a funky plot for other parameters.}

\item{threshold}{Logical. If TRUE (Default), will plot a dashed (upper) or dotted (lower) line if a water quality threshold exists for that

parameter and site. If FALSE, no threshold line will be plotted.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}

\item{facet_scales}{Specify whether facet axes should be "fixed" (all the same) or "free_y", "free_x" or "free" (both).}

\item{...}{Additional arguments relevant to \code{getChemistry()} or \code{getSondeInSitu()}}
}
\value{
Returns a ggplot object of specified current vs historic values
}
\description{
This function produces a plot that summarizes the range of historic data compared with current measurements.
Historic measurements are displayed as min-max values ever previously recorded (outermost band), upper and lower 95\%
distribution and middle 50\% distribution (inner quartiles) of values previously recorded (inner bands). The line
represents the median value.
}
\examples{
# Plot pH in Jordan Pond for 2023
plotWaterBands(site = "ACJORD", year_curr = 2023, years_historic = 2006:2022,
parameter = "pH", legend_position = 'right')

# Plot TN in Jordan Pond for 2023, including censored
plotWaterBands(site = "ACJORD", year_curr = 2023, years_historic = 2006:2022,

# Plot TN in Kroma Kill in SARA in 2023
plotWaterBands(site = "SARASA", year_curr = 2023, years_historic = 2006:2022, parameter = "TN",
legend_position = 'right')

}
