% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumEvents.R
\name{sumEvents}
\alias{sumEvents}
\title{sumEvents: summarize sample events}
\usage{
sumEvents(
  park = "all",
  site = "all",
  site_type = c("all", "lake", "stream"),
  event_type = "VS",
  years = 2006:format(Sys.Date(), "\%Y"),
  active = TRUE,
  months = 5:10
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{site}{Filter on 6-letter SiteCode (e.g., "ACABIN", "MORRSA", etc.). Easiest way to pick a site. Defaults to "all".}

\item{site_type}{Combine all site types, lakes or streams. Not needed if specifying particular sites.
\describe{
\item{"all"}{Default. Includes all site types, unless site or site_name select specific site types.}
\item{"lake"}{Include only lakes.}
\item{"stream"}{Include streams only.}
}}

\item{event_type}{Select the event type. Options available are below Can only choose one option.
\describe{
\item{"all"}{All possible sampling events.}
\item{"VS"}{Default. NETN Vital Signs monitoring events, which includes Projects named 'NETN_LS' and 'NETN+ACID'.}
\item{"acid"}{Acidification monitoring events in Acadia.}
\item{"misc"}{Miscellaneous sampling events.}
}}

\item{years}{Numeric. Years to query. Accepted values start at 2006.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites
that have been monitored.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 5 and 10, and these are set as the defaults.}

\item{parameter}{Specify the parameter(s) to return. Current accepted values are:.
chemistry: c("ANC_ueqL", "AppColor", "AppColor_PCU", "ChlA_ugL", "Cl_ueqL",
"DOC_mgL", "NH3_mgL", "NO2_mgL", "NO2+NO3_mgL", "NO3_ueqL", "pH_Lab", "PO4_ugL", "SO4_ueqL",
"TN_mgL", "TotDissN_mgL", "TotDissP_ugL", "TP_ugL")
sonde: c("Temp_C", "SpCond_uScm", "DOsat_pct", "DOsatLoc_pct", "DO_mgL", "pH", "pHmV",
"Turbidity_FNU", "ChlA_EXO_RFU", "ChlA_EXO_ugL", "BP_mmHg").
other: c("SDepth_m", "Discharge_cfs", "PenetrationRatio", "WaterLevel_Feet", "WaterLevel_m").
Note that "all" is the default value.}
}
\value{
Data frame of summarize event info
}
\description{
Summarize the number of samples collected at each site by year, month, parameter, and
whether measurment was real or censored.
}
\examples{
\dontrun{
importData()

# get all events for ACAD for all years and active sites
acad_ev <- sumEvents(park = "ACAD")

# get only lake events for ACAD for all years
acad_lk <- sumEvents(park = "ACAD", site_type = "lake")

# get LNETN events only
lnetn <- sumEvents(park = "LNETN")

}
}
